core: Move FontFace to font/font_face.rs
This makes font.rs smaller, as it's a pretty big file.
